+++ contracts/mocks/TestSyloUtils.sol
@@ -4,11 +4,11 @@ pragma solidity ^0.8.18;
 import "../libraries/SyloUtils.sol";
 
 contract TestSyloUtils {
-    function percOf(uint128 value, uint16 percentage) public pure returns (uint256) {
+    function percOf(uint128 value, uint32 percentage) public pure returns (uint256) {
         return SyloUtils.percOf(value, percentage);
     }
 
-    function asPerc(uint128 numerator, uint256 denominator) public pure returns (uint16) {
+    function asPerc(uint128 numerator, uint256 denominator) public pure returns (uint32) {
         return SyloUtils.asPerc(numerator, denominator);
     }
